pyproject.toml
[tool.poetry]
name = "qtranscribe"
version = "0.0.1"
description = "QTranscribe is a cloud service that uses Whisper model `medium` to provide in real-time transcription of audio files and youtube videos."
authors = ["obahamonde <oscarmartinbahamondemunoz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = "^20231117"
pytube = "12.1.3"
pydub = "^0.25.1"
numpy = "^1.26.4"
fastapi = "^0.111.0"
requests = "^2.32.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
